Refactor actions button into its own component
